chore: reorganize components and styles
