+++ site/src/App.tsx
@@ -1,20 +1,19 @@
-import './styles'
+import 'normalize.css'
+import './App.styl'
 import { StrictMode } from 'react'
 import styles from './Toast.module.styl'
 import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import { Header } from './components/Header'
-import { Footer } from './components/Footer'
 import { Main } from './components/Main'
+import { Footer } from './components/Footer'
 
 export const App = () => {
   return (
-    <>
-      <StrictMode>
-        <Header />
-        <Main />
-        <Footer />
-      </StrictMode>
+    <StrictMode>
+      <Header />
+      <Main />
+      <Footer />
 
       <ToastContainer
         className={styles.toast}
@@ -29,6 +28,6 @@ export const App = () => {
         draggable={false}
         pauseOnHover={false}
       />
-    </>
+    </StrictMode>
   )
 }
