+++ site/src/components/Header/components/ChangeLocale/ChangeLocale.module.styl
@@ -1,6 +1,9 @@
+@import '../../../../styles/settings/_variables'
+@import '../../../../styles/helpers/_vertical-rhythm'
+
 .box
   text-align: left
-  @media (min-width: em($tablet))
+  @media (min-width: $tablet)
     position absolute
 
 .button
