+++ site/src/components/Main/components/GenerateSection/GenerateSection.tsx
