+++ site/src/components/Main/components/InfoSection/InfoSection.module.styl
@@ -1,3 +1,6 @@
+@import '../../../../styles/settings/_variables'
+@import '../_center'
+
 .info-section
   background: $bg-color
