+++ site/src/components/Main/components/InfoSection/InfoSection.tsx
