+++ site/src/components/Main/components/InfoSection/index.ts
