+++ site/src/components/Main/components/ValidateSection/ValidateSection.tsx
