+++ site/src/components/Main/components/_center.styl
@@ -1,6 +1,3 @@
-$disable-select
-  user-select: none
-
 $center
   max-width: 1000px
   padding: 0 20px
