--- site/src/styles/demo.styl
@@ -1,59 +1,0 @@
-/*! ==================================================================
-   Main Styles
-   ================================================================== */
-
-/*!
-*
-* Box model reset
-* http://css-tricks.com/inheriting-box-sizing-probably-slightly-better-best-practice/
-*
-*/
-
-html
-  font-size: 100%
-  box-sizing: border-box
-
-*, *:before, *:after
-  box-sizing: inherit
-
-body
-  if $debug-vertical-rhythm == true
-    background-image: linear-gradient(to top, rgba($debug-color-line, 0.2) 1px, rgba(0, 0, 0, 0) 1px)
-    background-size: 100% em($base-line-height, $base-font-size)
-    background-position: left top
-
-  font: normal normal  $base-font-size\/em($base-line-height) $font-family-base
-  background-color: darken($bg-color, 30%)
-  color: $base-text-color
-  text-rendering: optimizeLegibility
-  -webkit-font-smoothing: antialiased
-  -moz-osx-font-smoothing: grayscale
-  font-smoothing: auto
-  font-feature-settings: "liga", "kern"
-
-/*!
-*
-* Better Font Smoothing in Chrome & Safari
-* http://www.intridea.com/blog/2014/5/8/better-font-smoothing-in-chrome-on-mac-os-x
-*
-*/
-@media only screen and (-webkit-min-device-pixel-ratio: 1.25),\
-  only screen and ( min-device-pixel-ratio: 1.25),\
-  only screen and ( min-resolution: 200dpi),\
-  only screen and ( min-resolution: 1.25dppx)
-
-  body
-    -webkit-font-smoothing: subpixel-antialiased
-
-::selection
-  background: $selection-bg-color
-  color: $selection-color
-
-@import "./partials/index.styl"
-
-
-/*! ==================================================================
-   Media Queries
-   ================================================================== */
-@media print
-    @import "./partials/print"
