--- site/src/styles/helpers/index.styl
@@ -1,3 +1,0 @@
-@import "./_functions"
-@import "./_placeholders"
-@import "./_vertical-rhythm"
