--- site/src/styles/index.ts
@@ -1,2 +1,0 @@
-import 'normalize.css'
-import './demo.styl'
