--- site/src/styles/partials/abbr.styl
@@ -1,2 +1,0 @@
-abbr
-    text-decoration: underline
