+++ site/src/styles/partials/blockquote.styl
@@ -1,3 +1,6 @@
+@import '../settings/_variables'
+@import '../helpers/_vertical-rhythm'
+
 blockquote
   font-style: italic
   padding-left: em(10px)
