--- site/src/styles/partials/button.styl
@@ -1,41 +1,0 @@
-$basicButton
-  adjust-font-size-to($base-font-size - 2)
-  border: none
-  border-radius: 5px
-  display: inline-block
-  padding: 0 em(20px)
-  overflow: hidden
-  margin: 0
-  text-align: center
-  text-decoration: none
-  text-transform: uppercase
-  user-select: none
-  vertical-align: bottom
-  -webkit-tap-highlight-color: transparent
-  box-shadow: 0 1px 2px rgba(0, 0, 0, .1), 0 1px 2px rgba(0, 0, 0, 0.1)
-  color: #fff
-  background-color: #3FAA96
-  transition: .15s ease-out
-  cursor: pointer
-
-  // &:focus
-  //  outline none
-  //  box-shadow 0 0 4px aqua
-
-button,
-input[type="button"]
-input[type="reset"]
-input[type="submit"]
-  @extend $basicButton
-
-  &:hover:not([disabled])
-    background-color: #2bbbad
-    box-shadow: 0 1px 4px rgba(0, 0, 0, .15), 0 1px 4px rgba(0, 0, 0, .15)
-
-  // &:active
-
-  &:disabled
-    @extend $basicButton
-    background-color: #DFDFDF
-    color: #9F9F9F
-    cursor: default
