+++ site/src/styles/partials/code.styl
@@ -1,7 +1,9 @@
+@import '../settings/_variables'
+@import '../helpers/_vertical-rhythm'
+
 code
   padding: 0 0.2em
   display: inline-block
-  border-radius: em(3px)
+  border-radius: 3px
   vertical-align: bottom
   overflow: auto
-  // rhythm-margins(1, 1, $base-font-size)
