--- site/src/styles/partials/fieldset.styl
@@ -1,4 +1,0 @@
-fieldset
-  margin: 0
-  rhythm-borders(1px, 1, $base-font-size, solid lighten(#000, 60%))
-  trailer(1)
