+++ site/src/styles/partials/index.styl
@@ -1,11 +1,11 @@
-@import 'headers';
-@import 'paragraph';
-@import 'lists';
+@import '../settings/_variables';
+@import '../helpers/_functions';
+@import '../helpers/_vertical-rhythm';
+
 @import 'blockquote';
-@import 'links';
-@import 'kbd';
 @import 'code';
-@import 'pre';
-@import 'fieldset';
+@import 'headers';
 @import 'input';
-@import 'textarea';
+@import 'links';
+@import 'paragraph';
+@import 'pre';
