+++ site/src/styles/partials/input.styl
@@ -1,3 +1,5 @@
+@import '../helpers/_functions'
+
 $input
   adjust-font-size-to($base-font-size)
   border: none
