--- site/src/styles/partials/kbd.styl
@@ -1,8 +1,0 @@
-kbd
-  color: #fff
-  font-family: $font-family-mono
-  background-color: rgba(0, 0, 0, .3)
-  border-radius: em(3px)
-  display: inline-block
-  vertical-align: bottom
-  padding: 0 0.2em 0 0.2em
