--- site/src/styles/partials/lists.styl
@@ -1,7 +1,0 @@
-ul,
-ol,
-dl
-  rhythm-margins(1, 1, $base-font-size)
-
-  li &
-    margin: 0
