+++ site/src/styles/partials/pre.styl
@@ -1,7 +1,10 @@
+@import '../settings/_variables'
+@import '../helpers/_vertical-rhythm'
+
 pre
   overflow: auto
   font-feature-settings: "tnum"
 
   code
-    padding: 0 em($base-line-height / 2)
+    padding: 0 $base-line-height / 2
     rhythm-padding(.5, .5, $base-font-size)
