--- site/src/styles/partials/textarea.styl
@@ -1,2 +1,0 @@
-textarea
-  @extend $input
