+++ site/src/styles/settings/_variables.styl
@@ -1,17 +1,13 @@
 // Color
 $primary-color = #217bb8
-$secondary-color = #000
-$tertiary-color = #000
 
 $bg-color = $primary-color
 $selection-color = #FFF
 $selection-bg-color = $primary-color
 
 // Typography
-$font-family-serif = "Times New Roman", Times, Georgia, Serif;
-$font-family-sans = "Inter", Helvetica, Arial, sans-serif
 $font-family-mono = "Courier New", Courier, monospace
-$font-family-base = $font-family-sans
+$font-family-base = "Inter", Helvetica, Arial, sans-serif
 
 $base-font-size = 16px
 $base-line-height = 24px
@@ -24,7 +20,6 @@ $font-size-h4 = 26px
 $font-size-h5 = 24px
 $font-size-h6 = 20px
 
-$header-weight = bold
 
 // Vertical Rhythmn
 $debug-vertical-rhythm = false
